+++ src/routing.c
@@ -43,7 +43,7 @@ int RouteMap_insert(RouteMap *routes, const char *pattern, size_t len, void *dat
     size_t pref_len = first_paren ? first_paren - pattern : len;
 
     route = RouteMap_insert_base(routes, pattern, pref_len, pattern, len);
-    check(route, "Failed to insert route: %.*s", len, pattern);
+    check(route, "Failed to insert route: %.*s", (int)len, pattern);
 
     route->data = data;
 
@@ -61,14 +61,14 @@ int RouteMap_insert_reversed(RouteMap *routes, const char *pattern, size_t len, 
     size_t prefix_len = last_paren ? len - (last_paren - pattern + 1) : len;
     int pi, ri;
 
-    debug("prefix_len: %d", prefix_len);
+    debug("prefix_len: %d", (int)prefix_len);
 
     for(pi = len-1, ri = 0; ri < prefix_len; ri++, pi--) {
         reversed_prefix[ri] = pattern[pi];
     }
     reversed_prefix[prefix_len] = '\0';
 
-    debug("prefix: %.*s, pattern: %.*s", prefix_len, reversed_prefix, len, pattern);
+    debug("prefix: %.*s, pattern: %.*s", (int)prefix_len, reversed_prefix, (int)len, pattern);
 
     route = RouteMap_insert_base(routes, reversed_prefix, prefix_len, pattern, len);
     check(route, "Failed to add host.");
